Added tag parser support for positional arguments.
1) The tag parser can now parse positional arguments:
{% latex "positional argument 1"%}...
The parser supports this, but code generation does not (yet).
2) the lmlexer now also supports positional arguments (necessary prereq
for 1)
3) Added more tests.
4) Changed the tokens so that they're "constructed" step by step out of
components. I'm not sure if this is easier or harder to read (see the
top of lmlexer.py)
5) Added flag to latex call so that it would fail immediately on error,
rather than prompting for user input.
6) Adjusted the regex for LSTART so that it could match more than just
'latex'. This opens the door for allowing more functions, other than
latex. Hopefully one day there will be a plugin system.
7) Shortened 'endlatex' on the LEND tag to just 'end'. In the future, if I add
more tags, they'll all use the same 'end' tag.