Started adding plugin functionality.
1) Started adding plugin functionality through tag_plugins. You can add
a function name (eg `latex`) and a reference to the constructor of the
class that can handle that function. (So latex points to
latexgen.LatexGen).
2) Finally plays nice with other tags that look the same (eg,
octopress plugins that also use {% %} to surround tags). Added a test to
test this.
3) In order to play nice with the other tags, I've added functionality
to the parser that builds a regex from the list of function names in
tag_plugins. So, other tags are simply ignored at the lexing stage. I'm
not sure if this is the best way to do it, but it was the easiest to
implement.