Rewrites lmparser to handle unary tags.
1) There was a issue with Unary tags and how they were handled in the
pervious version of the parser. The previous version expected the token
stream to always alternate between Latex tags (now BinTag) and Markdown
tags. Unary tags broke then invariant, so I had to write a real parser
instead of a cheap hack ;)
2) New parser is a stack based parser. Tags are pushed onto the stack
and popped off based on certain rules. By the time the end of the token
stream is reached, the stack is an AST. Pretty neat.
3) Updated tests to conform to the new parser. New ones still need to be
written to test the UnaryTag functionality.